rename customize-audio-node.js -> customize-audo-node-prototype.js
